5560hw2/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
matplotlib